fix some typescript errors; move hooks into subfolders
