+++ client/src/components/Today/Notes.tsx
@@ -3,8 +3,8 @@ import Empty from "@/components/Today/Empty";
 import Modal from "@/components/utility/Modal/Modal";
 import { filterTagsById } from "@/lib/filter-tags";
 import modalIds from "@/lib/modal-ids";
-import useNotesQuery from "@/lib/query/useNotesQuery";
-import useTagsQuery from "@/lib/query/useTagsQuery";
+import useNotesQuery from "@/lib/query/notes/useNotesQuery";
+import useTagsQuery from "@/lib/query/tags/useTagsQuery";
 import { useModalState } from "@/lib/state/modal-state";
 import { isToday } from "@lib/datetime/compare";
 import { Note } from "./Note";
