+++ client/src/components/auth/Register/useRegister.ts
@@ -1,4 +1,4 @@
-import useRegisterMutation from "@/lib/query/useRegisterMutation";
+import useRegisterMutation from "@/lib/query/user/useRegisterMutation";
 import type { NewUser } from "@type/server/user.types";
 import { useState } from "react";
