+++ client/src/lib/hooks/useAuthentication.ts
@@ -1,6 +1,6 @@
-import useLoginMutation from "../query/useLoginMutation";
-import useLogoutMutation from "../query/useLogoutMutation";
-import useMeQuery from "../query/useMeQuery";
+import useLoginMutation from "../query/user/useLoginMutation";
+import useLogoutMutation from "../query/user/useLogoutMutation";
+import useMeQuery from "../query/user/useMeQuery";
 
 export default function useAuthentication() {
 	const { mutate: login } = useLoginMutation();
