+++ client/src/lib/hooks/useReconcileSession.ts
@@ -1,5 +1,5 @@
 import { useEffect } from "react";
-import useMeQuery from "../query/useMeQuery";
+import useMeQuery from "../query/user/useMeQuery";
 import { localUser } from "../user-storage";
 import useAuthentication from "./useAuthentication";
