+++ client/src/lib/query/notes/useNewNoteMutation.ts
@@ -1,7 +1,7 @@
 import { useMutation } from "@tanstack/react-query";
 import type { NoteInput, NoteWithIds } from "@type/server/note.types";
-import { createRequestConfig } from "../fetch/create-request-config";
-import { makeAuthorizedUrl } from "../fetch/make-authorized-url";
+import { createRequestConfig } from "../../fetch/create-request-config";
+import { makeAuthorizedUrl } from "../../fetch/make-authorized-url";
 
 async function postNote({ note, tagIds }: NoteInput) {
 	const url = makeAuthorizedUrl("/data/note");
