+++ client/src/lib/query/notes/useNotesQuery.ts
@@ -1,7 +1,7 @@
 import type { NotesData } from "@/types/data.types";
 import { useQuery } from "@tanstack/react-query";
-import { makeAuthorizedUrl } from "../fetch/make-authorized-url";
-import { defaultQueryConfig } from "../query-client";
+import { makeAuthorizedUrl } from "../../fetch/make-authorized-url";
+import { defaultQueryConfig } from "../../query-client";
 
 async function getNotes() {
 	const url = makeAuthorizedUrl("/data/notes");
