+++ client/src/lib/query/user/useLoginMutation.ts
@@ -1,9 +1,9 @@
+import { createRequestConfig } from "@/lib/fetch/create-request-config";
+import { baseUrl } from "@/lib/fetch/fetch-constants";
+import { localUser } from "@/lib/user-storage";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import type { Data } from "@type/query.types";
 import type { User, UserLogin } from "@type/server/user.types";
-import { createRequestConfig } from "../fetch/create-request-config";
-import { baseUrl } from "../fetch/fetch-constants";
-import { localUser } from "../user-storage";
 
 async function postLogin(user: UserLogin) {
 	return (
