+++ client/src/lib/query/user/useLogoutMutation.ts
@@ -1,7 +1,7 @@
+import { createRequestConfig } from "@/lib/fetch/create-request-config";
+import { baseUrl } from "@/lib/fetch/fetch-constants";
+import { localUser } from "@/lib/user-storage";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { createRequestConfig } from "../fetch/create-request-config";
-import { baseUrl } from "../fetch/fetch-constants";
-import { localUser } from "../user-storage";
 
 async function postLogout() {
 	return (await fetch(`${baseUrl}/auth/logout`, createRequestConfig.post())).json();
