+++ client/src/lib/query/user/useRegisterMutation.ts
@@ -1,8 +1,8 @@
+import { createRequestConfig } from "@/lib/fetch/create-request-config";
+import { baseUrl } from "@/lib/fetch/fetch-constants";
 import { useMutation } from "@tanstack/react-query";
 import type { Data } from "@type/query.types";
 import type { NewUser, User } from "@type/server/user.types";
-import { createRequestConfig } from "../fetch/create-request-config";
-import { baseUrl } from "../fetch/fetch-constants";
 
 async function postRegister(newUser: NewUser) {
 	return (
